Got rid of duplication in audit pages
